Refactor: Dto 변환 로직 Util 패키지로 이동
