--- src/main/java/com/example/algoyai/util/DemoUtil.java
@@ -1,5 +1,0 @@
-package com.example.algoyai.util;
-
-public class DemoUtil {
-  //	todo
-}
